feat: Nix package + devShell & NixOS + Home-Manager modules
We use crane to provide incremental caching to avoid unnecessary rebuilds.

Because @project-gauntlet/tools is a git:// dependency in NPM, we must use
fetchNpmDeps with a hash unique to package-lock.json.

librusty_v8 builds quite slowly, so we fetch the binary bindings from the
project's GitHub release and provide a .#fetch-rusty-v8-hashes package to
get the correct hashes when v8 updates.

The modules are quite basic for now, but once declarative configuration is
supported in Gauntlet, these can be built out more.

Also, we had to include a patch in Cargo.toml for libffi-sys because its
crates.io release is outdated and breaks the nix sandbox.